number 转 int/double

+++ frameworks/kits/taihe/include/pixel_map_taihe.h
@@ -47,12 +47,12 @@ public:
     PixelMap CreateAlphaPixelmapSync();
     int32_t GetBytesNumberPerRow();
     int32_t GetPixelBytesNumber();
-    void ScaleSync(float x, float y);
-    void ScaleWithAntiAliasingSync(float x, float y, AntiAliasingLevel level);
+    void ScaleSync(double x, double y);
+    void ScaleWithAntiAliasingSync(double x, double y, AntiAliasingLevel level);
     void CropSync(ohos::multimedia::image::image::Region const& region);
-    void RotateSync(float angle);
+    void RotateSync(double angle);
     void FlipSync(bool horizontal, bool vertical);
-    void OpacitySync(float rate);
+    void OpacitySync(double rate);
     void SetMemoryNameSync(string_view name);
     void ReleaseSync();
     bool GetIsStrideAlignment();
